Patch

+++ features/dmManager.js
@@ -3,12 +3,16 @@ import { chnFetcher } from "../base/channels.js";
 import { colors } from "../base/colors.js";
 import { guildId, ids } from "../base/ids.js";
 import { dmChnSentMsgsLogger } from "../handlers/logger.js";
+import { patchnoteGenerator } from "../handlers/patchnoteGen.js";
 import { discord_activities } from "../handlers/sequelize.js";
+const lastestOpenedDM = new Map();
 export default (client) => {
     const dmChn = chnFetcher(ids.dmMsgsChnId);
     client.on("messageCreate", async (message) => {
         if (message.author.bot)
             return;
+        if (message.channelId === ids.patchGeneratorChnId)
+            return patchnoteGenerator(message);
         if (message.channel.type === ChannelType.DM) {
             if ((await message.channel.messages.fetch({ limit: 3 }))
                 .map((msg) => {
@@ -69,29 +73,37 @@ export default (client) => {
         else {
             discord_activities
                 .increment("messages", { by: 1, where: { authDatumDiscordId: message.author.id } })
-                .catch((e) => console.log(`Error during updating discordActivity for ${message.member.displayName}`, e));
+                .catch((e) => console.log(`[Error code: 1110] Error during updating discordActivity for ${message.member?.displayName}`, e));
         }
-        if (message.channel.id === dmChn.id && message.member?.permissions.has("Administrator") && message.guild && message.content.length > 18) {
-            const msgContent = message.content.trim().split(" ");
-            const userId = msgContent.shift();
-            const embedCheck = msgContent.pop();
-            const isEmbed = embedCheck === "embed" && msgContent.length !== 0 ? true : false;
-            const member = userId ? message.guild.members.cache.get(userId) : undefined;
-            if (member) {
-                const sendedMsg = await msgSend(isEmbed);
+        if (message.channel.id === dmChn.id && message.member?.permissions.has("Administrator") && message.guild) {
+            try {
+                await new Promise((res) => setTimeout(res, 30));
+                await message.fetch();
+            }
+            catch (error) {
+                return;
+            }
+            let msgContent = message.content.trim().replaceAll("\\n", "\n").replaceAll("\\*", "\n — ");
+            const isEmbed = msgContent.endsWith("embed") ? true : false;
+            const userId = msgContent.split(" ").shift();
+            let member = message.guild.members.cache.get(userId);
+            const lastDM = lastestOpenedDM.get(message.author.id);
+            member ? (msgContent = msgContent.slice(userId.length).trim()) : "";
+            !member && lastDM ? (member = lastDM) : "";
+            if ((!lastDM || lastDM.user.id !== member?.user.id) && member)
+                lastestOpenedDM.set(message.author.id, member);
+            isEmbed && msgContent.length > 5 ? (msgContent = msgContent.slice(0, -5).trim()) : "";
+            if (member !== undefined) {
+                const sendedMsg = await (isEmbed
+                    ? member.send({
+                        embeds: [new EmbedBuilder().setColor(colors.default).setDescription(msgContent)],
+                    })
+                    : member.send(msgContent));
                 message.delete();
                 dmChnSentMsgsLogger(member, sendedMsg.content.length > 0 ? sendedMsg.content : sendedMsg.embeds[0].description, sendedMsg.id);
-                async function msgSend(isEmbed) {
-                    if (isEmbed) {
-                        return member.send({
-                            embeds: [new EmbedBuilder().setColor(colors.default).setDescription(msgContent.join(" "))],
-                        });
-                    }
-                    else {
-                        msgContent.push(embedCheck);
-                        return member.send(msgContent.join(" "));
-                    }
-                }
+            }
+            else {
+                throw { name: "Ошибка. Пользователь не участник сервера", code: "[Error code: 1113]" };
             }
         }
     });
